extracted DiscussionPreview from ProjectContent
