+++ frontend/src/components/project/ProjectContent.js
@@ -1,4 +1,4 @@
-import { Button, Link, Typography } from "@material-ui/core";
+import { Button, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import ExpandLessIcon from "@material-ui/icons/ExpandLess";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
@@ -16,7 +16,7 @@ import MiniProfilePreview from "../profile/MiniProfilePreview";
 import Posts from "./../communication/Posts.js";
 import DateDisplay from "./../general/DateDisplay";
 import ProjectStatus from "./ProjectStatus";
-import UnfoldMoreIcon from "@material-ui/icons/UnfoldMore";
+import DiscussionPreview from "./DiscussionPreview";
 
 const MAX_DISPLAYED_DESCRIPTION_LENGTH = 500;
 
@@ -134,27 +134,6 @@ const useStyles = makeStyles((theme) => ({
   },
   finishedDate: {
     marginTop: theme.spacing(0.5),
-  },
-  linkDiscussionPreview: {
-    color: theme.palette.secondary.main,
-  },
-  discussionPreview: {
-    borderBottom: `1px solid ${theme.palette.grey[500]}`,
-    borderTop: `1px solid ${theme.palette.grey[500]}`,
-    marginBottom: theme.spacing(4),
-    paddingTop: theme.spacing(1),
-    paddingBottom: theme.spacing(1),
-  },
-  topSectionDiscussionPreview: {
-    display: "flex",
-    justifyContent: "space-between",
-  },
-  headingDiscussionPreview: {
-    fontWeight: "bold",
-    marginBottom: theme.spacing(0.5),
-  },
-  iconDiscussionPreview: {
-    marginRight: theme.spacing(2),
   },
 }));
 
@@ -390,34 +369,3 @@ function CollaborateContent({ project, texts }) {
   );
 }
 
-function DiscussionPreview({ latestParentComment, discussionTabLabel, locale, project }) {
-  const classes = useStyles();
-  return (
-    <>
-      <Link
-        href={`${getLocalePrefix(locale)}/projects/${project.url_slug}#comments`}
-        underline="none"
-        className={classes.linkDiscussionPreview}
-      >
-        <div className={classes.discussionPreview}>
-          <div className={classes.topSectionDiscussionPreview}>
-            <Typography
-              display="inline"
-              color="primary"
-              className={classes.headingDiscussionPreview}
-            >
-              {discussionTabLabel}
-            </Typography>
-            <UnfoldMoreIcon color="secondary" className={classes.iconDiscussionPreview} />
-          </div>
-          <Posts
-            posts={latestParentComment}
-            type="preview"
-            user={latestParentComment.author_user}
-            truncate={3}
-          />
-        </div>
-      </Link>
-    </>
-  );
-}
